requirements.txt
torch
torchvision
numpy
matplotlib
pillow
h5py
